requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
altair==4.1.0
appnope==0.1.3
arctic==1.79.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bleach==4.1.0
blinker==1.4
cachetools==4.2.4
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama
dataclasses==0.8
dateparser==1.1.1
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.2.1
entrypoints==0.4
enum-compat==0.0.3
frozenlist==1.2.0
gitdb==4.0.9
GitPython==3.1.18
idna==3.3
idna-ssl==1.1.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.17.2
Jinja2==3.0.3
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
lz4==3.1.10
MarkupSafe==2.0.1
mistune==0.8.4
mockextras==1.0.2
multidict==5.2.0
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.5
notebook==6.4.10
numpy==1.19.5
packaging==21.3
pandas==1.0.3
pandocfilters==1.5.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.8.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.19.4
ptyprocess==0.7.0
pyarrow==6.0.1
pycparser==2.21
pydeck==0.6.2
Pygments==2.12.0
pymongo==3.11.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.0
python-binance==1.0.16
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==23.0.0
regex==2022.3.2
requests==2.27.1
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
streamlit==1.9.2
tenacity==8.0.1
terminado==0.12.1
testpath==0.6.0
toml==0.10.2
toolz==0.11.2
tornado==6.1
tqdm
traitlets==4.3.3
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
ujson==4.3.0
urllib3==1.26.9
validators==0.19.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==9.1
widgetsnbextension==3.6.0
yarl==1.7.2
zipp==3.6.0
